moved getElementClassNames out to a helper file
