+++ src/scripts/helpers.js
@@ -67,3 +67,29 @@ export const Logger = {
     console.error(`%c${LOG_LABEL}`, 'color: #e74c3c', ...args);
   },
 };
+
+/**
+ * Retrieves and formats the class names of an element.
+ * Truncates the list if it exceeds the maximum display length.
+ *
+ * @param {HTMLElement} element - The DOM element whose class names are to be retrieved.
+ * @param {number} [maxLength=100] - The maximum length of the class names string.
+ * @returns {string} A formatted string of class names.
+ */
+export function getElementClassNames(element, maxLength = 100) {
+  const classAttribute = element.getAttribute('class');
+  // Handle cases where class attribute is null or not a string
+  if (!classAttribute || typeof classAttribute !== 'string') return '';
+
+  // Split class names by whitespace and filter out empty strings
+  const classNames = classAttribute.split(/\s+/).filter(Boolean);
+  let elementClasses = '';
+  if (classNames.length > 0) {
+    elementClasses = `.${classNames.join(' .')}`;
+    if (elementClasses.length > maxLength) {
+      elementClasses = elementClasses.substring(0, maxLength - 3) + '...';
+    }
+  }
+
+  return elementClasses;
+}
